pinctrl: intel: broxton: Allow UART 2 pins to be GPIO
[Taken from https://bitbucket.org/andy-shev/linux/branch/topic/uart/rpm]

The UART driver is going to support runtime PM. The problem is that that is
used for console and we have no other possibility to wake it up as switching
pins to GPIO and mark them as wake sources.

In this particular case split the pin group to input (RxD, CTS) pins and the
rest to be able to switch them to GPIO mode and back.

TODO: What about the rest of UARTs? Is this patch okay to pinctrl-broxton.c
maintainers?

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Wan Ahmad Zainie <wan.ahmad.zainie.wan.mohamad@intel.com>